requirements.txt
python-binance==1.0.15
pandas==1.3.0
streamlit==1.4.0